src/Backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-cra-helper==2.0.1
django-proxy==1.2.2
dnspython==2.0.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipwhois==1.2.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
maxminddb==2.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pydash==7.0.4
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
ripe.atlas.cousteau==2.0.0
ripe.atlas.sagan==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
websocket-client==1.3.3
yarl==1.9.2
